+++ test/module-factory.test.tsx
@@ -22,11 +22,11 @@ describe('test module factory', function () {
     const module3 = factory3.build();
     {
       const { result } = renderHook(() => module2.useActions());
-      expect(Object.keys(result.current)).toHaveLength(2); // $setState + $reset
+      expect(Object.keys(result.current)).toHaveLength(3); // $setState + setState + $reset
     }
     {
       const { result } = renderHook(() => module3.useActions());
-      expect(Object.keys(result.current)).toHaveLength(3); // $setState + $reset + hello
+      expect(Object.keys(result.current)).toHaveLength(4); // $setState + setState + $reset + hello
     }
   });
 });